// src/app/Core/Models/product.ts
export interface Product {

  comments: string;
  category: number;
  availability:  boolean,
  id: number;
  tig_id: number;
  price: number;
  price_on_sale: number;
  discount:number;
  sale: boolean,
  owner: string;
  unit: string;
  name: string;
  quantityInStock: number;
  quantity_sold:number;
  isEditing: boolean;
  //sale_date: Date;

}